Added LinkedHashMap to Fleet

+++ src/adapter/UpdateAuto.java
@@ -2,5 +2,5 @@ package adapter;
 
 public interface UpdateAuto {
 	public void updateOptionSetName(String modelName, String optSetname, String newName);
-	public void updateOptionPrice(String modelName, String optName, String name, float newPrice);
+	public void updateOptionPrice(String modelName, String optName, String name, double newPrice);
 }
